Ij sonarqube2 (#47)
* remove html

* remove html

* add logger

* rename Schopenhauer

* static final

* code smells client 1

* rename properties

* code smells server

* remove bugs

* code smells

* add logger to client

* constructors

* static changes

* fix all bugs

* small changes

* change packages names

* map static final

* consts in gameplay screen

* static changes

* static changes2

* add config file to client logger

+++ server/src/server/powers/AggroTaker.java
@@ -1,6 +1,6 @@
 package server.powers;
 
-import server.LogicMap.LogicMapHandler;
+import server.logicmap.LogicMapHandler;
 import server.players.Player;
 import server.User;
 import server.bullets.BulletList;
@@ -34,7 +34,7 @@ public class AggroTaker extends APower {
             double distance;
             for(AOpponent opponent : this.opponentList.getOpponents()){
                 distance = Math.sqrt((Math.abs(player.getCenterY() - opponent.getCenterY())) * (Math.abs(player.getCenterY() - opponent.getCenterY())) +
-                        (Math.abs(opponent.getCenterX() - player.getCenterX()) * (Math.abs(opponent.getCenterX() - player.getCenterX()))));
+                        (double)(Math.abs(opponent.getCenterX() - player.getCenterX()) * (Math.abs(opponent.getCenterX() - player.getCenterX()))));
                 if(distance < 2000){
                     opponent.setOpponentAI("Chaser");
                     opponent.setIdOfChasedPlayer(player.getUser().getName());
